Split pagination-focused tool tests into dedicated modules
