+++ AGENTS.md
@@ -57,9 +57,10 @@ mcp-server/
 │   └── tools/                  # Unit test modules for each tool implementation
 │       ├── address_tools/      # Tests for address-related MCP tools
 │       │   ├── test_get_address_info.py        # Tests for get_address_info
-│       │   ├── test_get_address_logs.py        # Tests for get_address_logs
-│       │   ├── test_get_tokens_by_address.py   # Tests for get_tokens_by_address
-│       │   └── test_nft_tokens_by_address.py   # Tests for nft_tokens_by_address
+│       │   ├── test_get_address_logs.py              # Tests for get_address_logs
+│       │   ├── test_get_tokens_by_address.py         # Tests for get_tokens_by_address
+│       │   ├── test_nft_tokens_by_address.py         # Tests for nft_tokens_by_address
+│       │   └── test_nft_tokens_by_address_pagination.py  # Pagination scenarios for nft_tokens_by_address
 │       ├── block_tools/        # Tests for block-related MCP tools
 │       │   ├── test_get_block_info.py          # Tests for get_block_info
 │       │   └── test_get_latest_block.py        # Tests for get_latest_block
@@ -71,12 +72,14 @@ mcp-server/
 │       │   ├── test_inspect_contract_code.py   # Tests for inspect_contract_code
 │       │   └── test_read_contract.py           # Tests for read_contract
 │       ├── transaction_tools/  # Tests for transaction-related MCP tools
-│       │   ├── test_get_token_transfers_by_address.py  # Tests for get_token_transfers_by_address
-│       │   ├── test_get_transaction_info.py    # Tests for get_transaction_info
-│       │   ├── test_get_transaction_logs.py    # Tests for get_transaction_logs
-│       │   ├── test_get_transactions_by_address.py  # Tests for get_transactions_by_address
-│       │   ├── test_helpers.py                 # Tests for transaction helper utilities
-│       │   └── test_transaction_summary.py     # Tests for transaction_summary
+│       │   ├── test_get_token_transfers_by_address.py      # Tests for get_token_transfers_by_address
+│       │   ├── test_get_transaction_info.py        # Tests for get_transaction_info
+│       │   ├── test_get_transaction_logs.py        # Tests for get_transaction_logs
+│       │   ├── test_get_transaction_logs_pagination.py  # Pagination-focused logs tests
+│       │   ├── test_get_transactions_by_address.py      # Tests for get_transactions_by_address
+│       │   ├── test_get_transactions_by_address_pagination.py  # Pagination-focused transaction tests
+│       │   ├── test_helpers.py                     # Tests for transaction helper utilities
+│       │   └── test_transaction_summary.py         # Tests for transaction_summary
 │       ├── test_common.py            # Tests for shared utility functions
 │       ├── test_common_truncate.py   # Tests for truncation helpers
 │       ├── test_decorators.py        # Tests for logging decorators
